add clients and EndpointConfig
