--- src/test/java/autotests/DuckFlyTest.java
@@ -1,92 +1,0 @@
-package autotests;
-
-import com.consol.citrus.TestCaseRunner;
-import com.consol.citrus.annotations.CitrusResource;
-import com.consol.citrus.annotations.CitrusTest;
-import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.testng.annotations.Optional;
-import org.testng.annotations.Test;
-
-import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
-
-public class DuckFlyTest extends TestNGCitrusSpringSupport {
-
-    @Test(description = "Существующий id с активными крыльями ")
-    @CitrusTest
-    public void flyActiveWings(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 0.15, "rubber", "quack", "ACTIVE");
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{\n" + "  \"message\": \"I am flying :)\"\n" + "}", HttpStatus.OK);
-        delete(runner, "${duckId}");
-    }
-    @Test(description = "Существующий id со связанными крыльями")
-    @CitrusTest
-    public void flyFixedWings(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 0.15, "rubber", "quack", "FIXED");
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{\n" + "  \"message\": \"I can not fly :C\"\n" + "}", HttpStatus.OK);
-        delete(runner, "${duckId}");
-    }
-    @Test(description = "Существующий id с крыльями в неопределенном состоянии ")
-    @CitrusTest
-    public void flyUndefinedWings(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 0.15, "rubber", "quack", "UNDEFINED");
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{\n" + "  \"message\": \"Wings are not detected :(\"\n" + "}", HttpStatus.OK);
-        delete(runner, "${duckId}");
-    }
-    public void duckFly(TestCaseRunner runner, String id) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .get("/api/duck/action/fly")
-                .queryParam("id", id));
-    }
-
-
-    public void validateResponse(TestCaseRunner runner, String responseMessage, HttpStatus httpStatus) {
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(httpStatus)
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE).body(responseMessage));
-    }
-
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .post("/api/duck/create")
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body("{\n" + "  \"color\": \"" + color + "\",\n"
-                        + "  \"height\": " + height + ",\n"
-                        + "  \"material\": \"" + material + "\",\n"
-                        + "  \"sound\": \"" + sound + "\",\n"
-                        + "  \"wingsState\": \"" + wingsState
-                        + "\"\n" + "}"));
-    }
-    public void delete (TestCaseRunner runner, String id){
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .delete("/api/duck/delete")
-                .queryParam("id", id));
-    }
-}
