--- src/test/java/autotests/DuckQuackTest.java
@@ -1,113 +1,0 @@
-package autotests;
-
-import com.consol.citrus.TestCaseRunner;
-import com.consol.citrus.annotations.CitrusResource;
-import com.consol.citrus.annotations.CitrusTest;
-import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.testng.annotations.Optional;
-import org.testng.annotations.Test;
-
-
-import java.util.concurrent.atomic.AtomicInteger;
-
-import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
-
-public class DuckQuackTest extends TestNGCitrusSpringSupport {
-
-    @Test(description = "Корректный чётный id , корректный звук ")
-    @CitrusTest
-    public void quackEvenId(@Optional @CitrusResource TestCaseRunner runner) {
-
-        AtomicInteger id = new AtomicInteger();
-
-        while (true) {
-            createDuck(runner, "yellow", 0.15, "rubber", "quack", "FIXED");
-            extractor(runner);
-            runner.$(a -> {
-                id.set(Integer.parseInt(a.getVariable("duckId")));
-            });
-
-            if (id.get() % 2 == 0)
-                break;
-        }
-
-        duckQuack(runner, "${duckId}","1","2");
-        validateResponse(runner, "{\n" + "  \"sound\": \"quack-quack\"\n" + "}", HttpStatus.OK);
-        delete(runner, "${duckId}");
-    }
-    @Test(description = "Корректный нечётный id, корректный звук ")
-    @CitrusTest
-    public void quackNotEvenId(@Optional @CitrusResource TestCaseRunner runner) {
-
-        AtomicInteger id = new AtomicInteger();
-
-        while (true) {
-            createDuck(runner, "yellow", 0.15, "rubber", "quack", "FIXED");
-            extractor(runner);
-            runner.$(a -> {
-                id.set(Integer.parseInt(a.getVariable("duckId")));
-            });
-
-            if (id.get() % 2 == 1)
-                break;
-        }
-
-        duckQuack(runner, "${duckId}","1","1");
-        validateResponse(runner, "{\n" + "  \"sound\": \"quack\"\n" + "}", HttpStatus.OK);
-        delete(runner,"${duckId}" );
-    }
-
-
-    public static void extractor(TestCaseRunner runner) {
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
-
-    }
-
-
-    public void duckQuack(TestCaseRunner runner, String id,String repetitionCount, String soundCount) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .get("/api/duck/action/quack")
-                .queryParam("id", id)
-        .queryParam("repetitionCount", repetitionCount)
-                .queryParam("soundCount",soundCount));
-    }
-
-
-    public void validateResponse(TestCaseRunner runner, String responseMessage, HttpStatus httpStatus) {
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(httpStatus)
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE).body(responseMessage));
-    }
-
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .post("/api/duck/create")
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body("{\n" + "  \"color\": \"" + color + "\",\n"
-                        + "  \"height\": " + height + ",\n"
-                        + "  \"material\": \"" + material + "\",\n"
-                        + "  \"sound\": \"" + sound + "\",\n"
-                        + "  \"wingsState\": \"" + wingsState
-                        + "\"\n" + "}"));
-    }
-
-    public void delete(TestCaseRunner runner, String id) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .delete("/api/duck/delete")
-                .queryParam("id", id));
-    }
-}
