--- src/test/java/autotests/DuckUpdateTest.java
@@ -1,93 +1,0 @@
-package autotests;
-
-import com.consol.citrus.TestCaseRunner;
-import com.consol.citrus.annotations.CitrusResource;
-import com.consol.citrus.annotations.CitrusTest;
-import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.testng.annotations.Optional;
-import org.testng.annotations.Test;
-
-import java.util.concurrent.atomic.AtomicInteger;
-
-import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
-
-public class DuckUpdateTest extends TestNGCitrusSpringSupport {
-
-    @Test(description = "Изменяем цвет и высоту утки ")
-    @CitrusTest
-    public void updateColorAndHeightDuck(@Optional @CitrusResource TestCaseRunner runner) {
-        AtomicInteger id = new AtomicInteger();
-        createDuck(runner, "yellow", 0.15, "rubber", "quack", "FIXED");
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
-        runner.$(a -> {
-            id.set(Integer.parseInt(a.getVariable("duckId")));
-        });
-        duckUpdate(runner, "${duckId}", "10", "green", "rubber", "quack");
-        validateResponse(runner, "{\n" + "  \"message\": \"Duck with id = " +Integer.toString(id.get())+" is updated\"\n" + "}", HttpStatus.OK);
-        delete(runner, "${duckId}");
-    }
-
-    @Test(description = "Изменяем цвет и звук утки ")
-    @CitrusTest
-    public void updateColorAndSoundDuck(@Optional @CitrusResource TestCaseRunner runner) {
-
-        createDuck(runner, "yellow", 0.15, "rubber", "quack", "FIXED");
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
-
-        duckUpdate(runner, "${duckId}", "10", "yellow", "rubber", "meow");
-        validateResponse(runner, "{\n" + "@ignore@" + "}", HttpStatus.BAD_REQUEST);
-        delete(runner, "${duckId}");
-    }
-    public void duckUpdate(TestCaseRunner runner, String id, String height, String color, String material, String sound) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .put("/api/duck/update")
-                .queryParam("color", color)
-                .queryParam("height", height)
-                .queryParam("id", id)
-                .queryParam("material", material)
-                .queryParam("sound", sound));
-
-    }
-
-
-    public void validateResponse(TestCaseRunner runner, String responseMessage, HttpStatus httpStatus) {
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(httpStatus)
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE).body(responseMessage));
-    }
-
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .post("/api/duck/create")
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body("{\n" + "  \"color\": \"" + color + "\",\n"
-                        + "  \"height\": " + height + ",\n"
-                        + "  \"material\": \"" + material + "\",\n"
-                        + "  \"sound\": \"" + sound + "\",\n"
-                        + "  \"wingsState\": \"" + wingsState
-                        + "\"\n" + "}"));
-    }
-    public void delete (TestCaseRunner runner, String id){
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .delete("/api/duck/delete")
-                .queryParam("id", id));
-    }
-}
