+++ src/test/java/autotests/clients/DuckActionsClient.java
@@ -1,0 +1,106 @@
+package autotests.clients;
+
+import autotests.EndpointConfig;
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.beans.factory.annotation.Autowired;
+
+
+import com.consol.citrus.http.client.HttpClient;
+
+import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
+import static com.consol.citrus.http.actions.HttpActionBuilder.http;
+
+@ContextConfiguration(classes = {EndpointConfig.class})
+public class DuckActionsClient extends TestNGCitrusSpringSupport {
+
+    @Autowired
+    protected HttpClient yellowDuckService;
+
+
+    public void extractor(TestCaseRunner runner) {
+        runner.$(http().client(yellowDuckService)
+                .receive()
+                .response(HttpStatus.OK)
+                .message()
+                .extract(fromBody().expression("$.id", "duckId")));
+
+    }
+
+
+    public void duckProperties(TestCaseRunner runner, String id) {
+        runner.$(http().client(yellowDuckService)
+                .send()
+                .get("/api/duck/action/properties")
+                .queryParam("id", id));
+    }
+
+
+    public void validateResponse(TestCaseRunner runner, String responseMessage, HttpStatus httpStatus) {
+        runner.$(http().client(yellowDuckService)
+                .receive()
+                .response(httpStatus)
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE).body(responseMessage));
+    }
+
+
+    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
+        runner.$(http().client(yellowDuckService)
+                .send()
+                .post("/api/duck/create")
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body("{\n" + "  \"color\": \"" + color + "\",\n"
+                        + "  \"height\": " + height + ",\n"
+                        + "  \"material\": \"" + material + "\",\n"
+                        + "  \"sound\": \"" + sound + "\",\n"
+                        + "  \"wingsState\": \"" + wingsState
+                        + "\"\n" + "}"));
+    }
+
+    public void delete(TestCaseRunner runner, String id) {
+        runner.$(http().client(yellowDuckService)
+                .send()
+                .delete("/api/duck/delete")
+                .queryParam("id", id));
+    }
+
+    public void duckQuack(TestCaseRunner runner, String id, String repetitionCount, String soundCount) {
+        runner.$(http().client(yellowDuckService)
+                .send()
+                .get("/api/duck/action/quack")
+                .queryParam("id", id)
+                .queryParam("repetitionCount", repetitionCount)
+                .queryParam("soundCount", soundCount));
+    }
+
+    public void duckFly(TestCaseRunner runner, String id) {
+        runner.$(http().client(yellowDuckService)
+                .send()
+                .get("/api/duck/action/fly")
+                .queryParam("id", id));
+    }
+
+    public void duckSwim(TestCaseRunner runner, String id) {
+        runner.$(http().client(yellowDuckService)
+                .send()
+                .get("/api/duck/action/swim")
+                .queryParam("id", id));
+    }
+
+    public void duckUpdate(TestCaseRunner runner, String id, String height, String color, String material, String sound) {
+        runner.$(http().client(yellowDuckService)
+                .send()
+                .put("/api/duck/update")
+                .queryParam("color", color)
+                .queryParam("height", height)
+                .queryParam("id", id)
+                .queryParam("material", material)
+                .queryParam("sound", sound));
+
+    }
+}
