+++ src/test/java/autotests/tests/DuckCreateTest.java
@@ -1,0 +1,45 @@
+package autotests.tests;
+import autotests.clients.DuckActionsClient;
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.annotations.CitrusResource;
+import com.consol.citrus.annotations.CitrusTest;
+import org.springframework.http.HttpStatus;
+import org.testng.annotations.Optional;
+import org.testng.annotations.Test;
+
+
+public class DuckCreateTest extends DuckActionsClient {
+
+    @Test(description = "Создание утки с material = rubber ")
+    @CitrusTest
+    public void createMaterialRubber(@Optional @CitrusResource TestCaseRunner runner) {
+        createDuck(runner, "yellow", 0.15, "rubber", "quack", "ACTIVE");
+
+
+        validateResponse(runner, "{\n"
+                + "  \"color\": \"" + "yellow" + "\",\n"
+                + "  \"height\": " + 0.15 + ",\n"
+                + "  \"material\": \"" + "rubber" + "\",\n"
+                + "  \"sound\": \"" + "quack" + "\",\n"
+                + "  \"wingsState\": \"" + "ACTIVE"
+                + "\"\n" + "}", HttpStatus.OK);
+
+    }
+    @Test(description = "Создание утки с material = wood ")
+    @CitrusTest
+    public void createMaterialWood(@Optional @CitrusResource TestCaseRunner runner) {
+        createDuck(runner, "yellow", 0.15, "wood", "quack", "FIXED");
+
+
+        validateResponse(runner, "{\n"
+                + "  \"color\": \"" + "yellow" + "\",\n"
+                + "  \"height\": " + 0.15 + ",\n"
+                + "  \"material\": \"" + "wood" + "\",\n"
+                + "  \"sound\": \"" + "quack" + "\",\n"
+                + "  \"wingsState\": \"" + "FIXED"
+                + "\"\n" + "}", HttpStatus.OK);
+
+    }
+
+
+}
