+++ src/test/java/autotests/tests/DuckDeleteTest.java
@@ -1,0 +1,22 @@
+package autotests.tests;
+
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.annotations.CitrusResource;
+import com.consol.citrus.annotations.CitrusTest;
+import org.springframework.http.HttpStatus;
+import org.testng.annotations.Optional;
+import org.testng.annotations.Test;
+import autotests.clients.DuckActionsClient;
+
+public class DuckDeleteTest extends DuckActionsClient {
+
+    @Test(description = "Удаление существующей утки ")
+    @CitrusTest
+    public void deleteDuck(@Optional @CitrusResource TestCaseRunner runner) {
+        createDuck(runner, "yellow", 0.15, "rubber", "quack", "FIXED");
+        extractor(runner);
+        delete(runner, "${duckId}");
+        validateResponse(runner, "{\n" + "  \"message\": \"Duck is deleted\"\n" + "}", HttpStatus.OK);
+    }
+
+}
