+++ src/test/java/autotests/tests/DuckPropertiesTest.java
@@ -1,21 +1,17 @@
-package autotests;
+package autotests.tests;
 
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
-import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
-
+import autotests.clients.DuckActionsClient;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
-import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
-import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
-public class DuckPropertiesTest extends TestNGCitrusSpringSupport {
+public class DuckPropertiesTest extends DuckActionsClient {
 
     @Test(description = " ID - целое четное число. Материал wood ")
     @CitrusTest
@@ -73,51 +69,5 @@ public class DuckPropertiesTest extends TestNGCitrusSpringSupport {
     }
 
 
-    public static void extractor(TestCaseRunner runner) {
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(HttpStatus.OK)
-                .message()
-                .extract(fromBody().expression("$.id", "duckId")));
 
-    }
-
-
-    public void duckProperties(TestCaseRunner runner, String id) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .get("/api/duck/action/properties")
-                .queryParam("id", id));
-    }
-
-
-    public void validateResponse(TestCaseRunner runner, String responseMessage, HttpStatus httpStatus) {
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response(httpStatus)
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE).body(responseMessage));
-    }
-
-
-    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .post("/api/duck/create")
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body("{\n" + "  \"color\": \"" + color + "\",\n"
-                        + "  \"height\": " + height + ",\n"
-                        + "  \"material\": \"" + material + "\",\n"
-                        + "  \"sound\": \"" + sound + "\",\n"
-                        + "  \"wingsState\": \"" + wingsState
-                        + "\"\n" + "}"));
-    }
-
-    public void delete(TestCaseRunner runner, String id) {
-        runner.$(http().client("http://localhost:2222")
-                .send()
-                .delete("/api/duck/delete")
-                .queryParam("id", id));
-    }
 }
